#31 4K (not scaled) improvements.

+++ src/app/components/index/index.component.ts
@@ -5,6 +5,8 @@ import { EventRequest } from 'src/app/entity/EventRequest';
 import { CompetitionService } from 'src/app/service/competition.service';
 import { EventService } from 'src/app/service/event.service';
 import { TagsService } from 'src/app/service/tags.service';
+import {ConfirmDialogComponent} from "../../shared/components/confirm-dialog/confirm-dialog.component";
+import {MatDialog} from "@angular/material/dialog";
 
 @Component({
   selector: 'app-index',
@@ -30,6 +32,7 @@ export class IndexComponent implements OnInit {
     protected eventService: EventService,
     protected tagsService: TagsService,
     private router: Router,
+    private dialog: MatDialog,
   ) {}
 
   ngOnInit(): void {
@@ -110,6 +113,7 @@ export class IndexComponent implements OnInit {
 
     this.selectCompetitions();
     this.selectEvents();
+    if(this.tagsService.getAvailableTags().length === 0) this.closeTagSelector();
   }
 
   removeTag(tagIndex: number) {
@@ -144,12 +148,25 @@ export class IndexComponent implements OnInit {
     this.router.navigate(['/competitions/edit/', competitionId]);
   }
 
+
+
   handleCompetitionDeleteClick(competitionId: number) {
-    this.competitionService.deleteCompetition(competitionId).subscribe(() => {
-      this.competitionService.getCompetitions().subscribe((competitions) => {
-        this.competitions = competitions;
-        this.selectCompetitions();
-      });
+    const dialogRef = this.dialog.open(ConfirmDialogComponent, {
+      data: {
+        title: 'Delete competition',
+        content:
+          'Are you sure you want to delete this competition?',
+      },
+    });
+    dialogRef.afterClosed().subscribe((result) => {
+      if (result) {
+        this.competitionService.deleteCompetition(competitionId).subscribe(() => {
+          this.competitionService.getCompetitions().subscribe((competitions) => {
+            this.competitions = competitions;
+            this.selectCompetitions();
+          });
+        });
+      }
     });
   }
 }
